Refactor the physical file locations to add more structure to the project and reflect the VS filters
