+++ .gitignore
@@ -1,4 +1,4 @@
 #Framework Index
 #Ignore Visual Studio Files
-/.vs/
-/Debug/
+.vs/
+Debug/
