--- Framework.h
@@ -1,3 +1,0 @@
-#pragma once
-#include "Included.h"
-#include "Software.h"
