--- Hardware.h
@@ -1,3 +1,0 @@
-#pragma once
-#include "Mouse.h"
-#include "Keyboard.h"
