--- Interface.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "DirectX9.h"
-#include "GDI.h"
-#include "Window.h"
