--- Memory.h
@@ -1,3 +1,0 @@
-#pragma once
-#include "VMT.h"
-#include "Module.h"
