+++ Src/Framework.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "Software/Included.h"
+#include "Software/Libraries/Software.h"
