+++ Src/Software/Interface/DirectX9.cpp
