+++ Src/Software/Interface/DirectX9.h
@@ -1,7 +1,7 @@
 #pragma once
-#include "Included.h"
+#include "../Included.h"
 #include "Window.h"
-#include "Mouse.h"
+#include "../Hardware/Mouse.h"
 
 namespace Software 
 {
