+++ Src/Software/Libraries/Hardware.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "../Hardware/Mouse.h"
+#include "../Hardware/Keyboard.h"
