+++ Src/Software/Libraries/Interface.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "../Interface/DirectX9.h"
+#include "../Interface/GDI.h"
+#include "../Interface/Window.h"
