+++ Src/Software/Libraries/Memory.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "../Memory/Module.h"
+#include "../Memory/VMT.h"
